Create APLconceptualisation.md
This Programming Language is in developement. As a beginner who started learning Programming. I notice how overcomplicated the programming Languages are. It could be so simple and yes i understood the purpose why they are structured; 
1. Because of older Programming Languages. 
2. To avoid Errors, but that can be done with different Techniques today. 
I am proposing a new approach to make it easier and lightweighted. 
There is a lot of restructering and addition to add to the Alphabetic Programming Language. But i am aware of it and i will do it some day or the community will help me out. As i hope so.